Computes technical coefficients matrix.

Computes technical coefficients matrix.

compute_tech_coeff(intermediate_transactions, total_production)

Arguments

intermediate_transactions

A \(n x n\) matrix of intermediate transactions.

total_production

A \(1 x n\) vector of total production.

Value

A \(n x n\) matrix of technical coefficients, known as A matrix.

Details

It calculates the technical coefficients matrix, which is the columnwise ratio of intermediate transactions to total production (Leontief 1983) .

References

Leontief W (1983). A Economia do Insumo-Produto, Os Economistas. Abril Cultural, São Paulo.

Underlined Rust code uses Rayon crate to parallelize the computation by default, so there is no need to use future or async/await to parallelize.

Examples

intermediate_transactions <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), 3, 3)
total_production <- matrix(c(100, 200, 300), 1, 3)
# instantiate iom object
my_iom <- fio::iom$new("test", intermediate_transactions, total_production)
# Calculate the technical coefficients
my_iom$compute_tech_coeff()
# show the technical coefficients
my_iom$technical_coefficients_matrix
#>      1     2          3
#> 1 0.01 0.020 0.02333333
#> 2 0.02 0.025 0.02666667
#> 3 0.03 0.030 0.03000000