R6 class for input-output matrix

R6 class for input-output matrix.

Public fields

id

Identifier of the new instance

intermediate_transactions

Intermediate transactions matrix.

total_production

Total production vector.

household_consumption

Household consumption vector.

government_consumption

Government consumption vector.

exports

Exports vector.

final_demand_others

Other vectors of final demand that doesn't have dedicated slots.

final_demand_matrix

Aggregates final demand vectors into a matrix.

imports

Imports vector.

taxes

Taxes vector.

wages

Wages vector.

operating_income

Operating income vector.

added_value_others

Other vectors of added value that doesn't have dedicated slots.

added_value_matrix

Aggregates added value vectors into a matrix.

occupation

Occupation vector

technical_coefficients_matrix

Technical coefficients matrix.

leontief_inverse_matrix

Leontief inverse matrix.

multiplier_output

Output multiplier dataframe.

multiplier_employment

Employment multiplier dataframe.

multiplier_taxes

Taxes multiplier dataframe.

multiplier_wages

Wages multiplier dataframe.

field_influence

Influence field matrix.

key_sectors

Key sectors dataframe.

allocation_coefficients_matrix

Allocation coefficients matrix.

ghosh_inverse_matrix

Ghosh inverse matrix.

hypothetical_extraction

Absolute and relative backward and forward differences in total output after a hypothetical extraction.

threads

Number of threads available for Rust to run in parallel. Defaults to 0, meaning running in parallel with all available threads.

Methods

Public methods

iom$new()

iom$add()

iom$remove()

iom$update_final_demand_matrix()

iom$update_added_value_matrix()

iom$compute_tech_coeff()

iom$compute_leontief_inverse()

iom$compute_multiplier_output()

iom$compute_multiplier_employment()

iom$compute_multiplier_wages()

iom$compute_multiplier_taxes()

iom$compute_field_influence()

iom$compute_key_sectors()

iom$compute_allocation_coeff()

iom$compute_ghosh_inverse()

iom$compute_hypothetical_extraction()

iom$set_max_threads()

iom$clone()

Method new()

Creates a new instance of this R6 class.

Usage

iom$new(
  id,
  intermediate_transactions,
  total_production,
  household_consumption = NULL,
  government_consumption = NULL,
  exports = NULL,
  final_demand_others = NULL,
  imports = NULL,
  taxes = NULL,
  wages = NULL,
  operating_income = NULL,
  added_value_others = NULL,
  occupation = NULL
)

Arguments

id

Identifier for the input-output matrix.

intermediate_transactions

Intermediate transactions matrix.

total_production

Total production vector.

household_consumption

Household consumption vector.

government_consumption

Government consumption vector.

exports

Exports vector.

final_demand_others

Other vectors of final demand that doesn't have dedicated slots. Setting column names is advised for better readability.

imports

Imports vector.

taxes

Taxes vector.

wages

Wages vector.

operating_income

Operating income vector.

added_value_others

Other vectors of added value that doesn't have dedicated slots. Setting row names is advised for better readability.

occupation

Occupation matrix.

Method add()

Adds a matrix to a previously imported IO matrix.

Usage

iom$add(matrix_name, matrix)

Arguments

matrix_name

One of household_consumption, government_consumption, exports, final_demand_others, imports, taxes, wages, operating income, added_value_others or occupation matrix to be added.

matrix

Matrix object to be added.

Method remove()

Removes a matrix from a previously imported IO matrix.

Usage

iom$remove(matrix_name)

Arguments

matrix_name

One of household_consumption, government_consumption, exports, final_demand_others, imports, taxes, wages, operating income, added_value_others or occupation matrix to be removed.

matrix

Matrix object to be removed.

Method update_final_demand_matrix()

Updates final demand matrix.

Usage

iom$update_final_demand_matrix()

Method update_added_value_matrix()

Updates added value matrix.

Usage

iom$update_added_value_matrix()

Method compute_tech_coeff()

Computes the technical coefficients matrix.

Usage

iom$compute_tech_coeff()

Method compute_leontief_inverse()

Computes the Leontief inverse matrix.

Usage

iom$compute_leontief_inverse()

Arguments

technical_coefficients_matrix

Technical coefficients matrix.

Method compute_multiplier_output()

Computes the output multiplier dataframe.

Usage

iom$compute_multiplier_output()

Arguments

leontief_inverse_matrix

Leontief inverse matrix.

Method compute_multiplier_employment()

Computes the employment multiplier dataframe.

Usage

iom$compute_multiplier_employment()

Arguments

leontief_inverse_matrix

Leontief inverse matrix.

Method compute_multiplier_wages()

Computes the wages multiplier dataframe.

Usage

iom$compute_multiplier_wages()

Arguments

leontief_inverse_matrix

Leontief inverse matrix.

Method compute_multiplier_taxes()

Computes the taxes multiplier dataframe.

Usage

iom$compute_multiplier_taxes()

Arguments

leontief_inverse_matrix

Leontief inverse matrix.

Method compute_field_influence()

Computes the influence field matrix.

Usage

iom$compute_field_influence(epsilon)

Arguments

epsilon

Epsilon value. A technical change in the input-output matrix.

Method compute_key_sectors()

Computes the key sectors dataframe.

Usage

iom$compute_key_sectors()

Method compute_allocation_coeff()

Computes the allocation coefficients matrix.

Usage

iom$compute_allocation_coeff()

Method compute_ghosh_inverse()

Computes the Ghosh inverse matrix.

Usage

iom$compute_ghosh_inverse()

Method compute_hypothetical_extraction()

Computes the hypothetical extraction.

Usage

iom$compute_hypothetical_extraction()

Method set_max_threads()

Sets max number of threads used by fio.

Usage

iom$set_max_threads(max_threads)

Arguments

max_threads

Number of threads enabled for parallel computing. Defaults to the number of threads available.

Method clone()

The objects of this class are cloneable with this method.

Usage

iom$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# data
intermediate_transactions <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), 3, 3)
total_production <- matrix(c(100, 200, 300), 1, 3)
exports <- matrix(c(10, 20, 30), 3, 1)
households <- matrix(4:6, 3, 1)
imports <- matrix(c(5, 10, 15), 1, 3)
jobs <- matrix(c(10, 12, 15), 1, 3)
taxes <- matrix(c(2, 5, 10), 1, 3)
wages <- matrix(c(11, 12, 13), 1, 3)

# a new iom instance can be created by passing just intermediate transactions and total production
my_iom <- iom$new(
 "example",
 intermediate_transactions,
 total_production
)

# Compute technical coefficients matrix
my_iom$compute_tech_coeff()
print(my_iom$technical_coefficients_matrix)
#>      1     2          3
#> 1 0.01 0.020 0.02333333
#> 2 0.02 0.025 0.02666667
#> 3 0.03 0.030 0.03000000

# Compute leontief inverse matrix
my_iom$compute_leontief_inverse()
print(my_iom$leontief_inverse_matrix)
#>            1          2          3
#> 1 1.01129067 0.02151113 0.02491795
#> 2 0.02161815 1.02696918 0.02875285
#> 3 0.03194563 0.03242723 1.03258776

# compute output multiplier
my_iom$compute_multiplier_output()
print(my_iom$multiplier_output)
#>   sector multiplier_total multiplier_direct multiplier_indirect
#> 1      1         1.064854             0.060            1.004854
#> 2      2         1.080908             0.075            1.005908
#> 3      3         1.086259             0.080            1.006259

# `add` or `remove` other elements to IO matrix
my_iom$add("exports", exports)
my_iom$add("household_consumption", households)
my_iom$add("occupation", jobs)
my_iom$add("taxes", taxes)
my_iom$add("wages", wages)

# compute multipliers for added elements
my_iom$compute_multiplier_employment()
my_iom$compute_multiplier_taxes()
my_iom$compute_multiplier_wages()
print(my_iom$multiplier_wages)
#>   sector multiplier_total multiplier_direct multiplier_indirect
#> 1      1       0.11392337        0.11000000         0.003923373
#> 2      2       0.06538955        0.06000000         0.005389555
#> 3      3       0.04921162        0.04333333         0.005878282

# compute field of influence
my_iom$compute_field_influence(epsilon = 0.001)

# compute power and sensitivity of dispersion, it's coefficients of variation
# and identify key sectors
my_iom$compute_key_sectors()
print(my_iom$key_sectors)
#>   sector power_dispersion sensitivity_dispersion power_dispersion_cv
#> 1      1        0.9884106              0.9817881            1.617480
#> 2      2        1.0033113              1.0000000            1.602404
#> 3      3        1.0082781              1.0182119            1.588209
#>   sensitivity_dispersion_cv             key_sectors
#> 1                  1.628303          Non-Key Sector
#> 2                  1.607718 Strong Backward Linkage
#> 3                  1.572583              Key Sector

# supply-wise model
my_iom$compute_allocation_coeff()
my_iom$compute_ghosh_inverse()
print(my_iom$ghosh_inverse_matrix)
#>            1          2          3
#> 1 1.01129067 0.04302226 0.07475385
#> 2 0.01080908 1.02696918 0.04312928
#> 3 0.01064854 0.02161815 1.03258776

# aggregates final demand and added value vectors
my_iom$update_final_demand_matrix()
print(my_iom$final_demand_matrix)
#>      [,1] [,2]
#> [1,]    4   10
#> [2,]    5   20
#> [3,]    6   30

my_iom$update_added_value_matrix()
print(my_iom$added_value_matrix)
#>      [,1] [,2] [,3]
#> [1,]    2    5   10
#> [2,]   11   12   13

# perform hypothetical extraction of a given sector
my_iom$compute_hypothetical_extraction()
print(my_iom$hypothetical_extraction)
#>   backward_absolute backward_relative forward_absolute forward_relative
#> 1         -519.9640        -0.8666067        -544.1876       -0.9069794
#> 2         -521.0921        -0.8684869        -543.9137       -0.9065229
#> 3         -522.1744        -0.8702907        -544.0538       -0.9067563
#>   total_absolute total_relative
#> 1      -1064.152      -1.773586
#> 2      -1065.006      -1.775010
#> 3      -1066.228      -1.777047